# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Project Kessan API
  description: |
    AI-Powered Japanese Stock Trend Analysis Platform
    
    Project Kessan (決算 - "Financial Results") provides institutional-grade investment analysis 
    for Japanese stocks through AI-powered insights, real-time data, and comprehensive financial analysis.
    
    ## Features
    - **AI Analysis**: Multi-horizon stock analysis using Google Gemini API
    - **Real-time Data**: Live stock prices and market indices
    - **Japanese Focus**: Specialized for Tokyo Stock Exchange (TSE) listed companies
    - **News & Sentiment**: Japanese language news analysis with sentiment scoring
    - **Subscription Tiers**: Flexible pricing with quota management
    - **Watchlist Management**: Personal stock tracking and monitoring
    
    ## Authentication
    The API uses JWT (JSON Web Tokens) for authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limits
    - **Free Tier**: 20 requests/minute (anonymous), 100 requests/minute (authenticated)
    - **Pro Tier**: 300 requests/minute
    - **Business Tier**: 1000 requests/minute
    
    ## Error Handling
    The API returns structured error responses with appropriate HTTP status codes:
    - `400` - Bad Request (validation errors)
    - `401` - Unauthorized (authentication required)
    - `403` - Forbidden (insufficient permissions)
    - `404` - Not Found (resource not found)
...
    
...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /subscription/plans:
    get:
      tags:
        - Subscription
      summary: Get subscription plans
      description: Get all available subscription plans with features and pricing
      security: []
      parameters:
        - name: active_only
          in: query
          schema:
            type: boolean
            default: true
          description: Return only active plans
      responses:
        '200':
          description: Available subscription plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Plan'
  /subscription/my-subscription:
    get:
      tags:
        - Subscription
      summary: Get user subscription
      description: Get current user's subscription information with usage data
      responses:
        '200':
          description: User subscription details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionWithUsage'
        '401':
          description: Authentication required
  /subscription/upgrade:
    post:
      tags:
        - Subscription
      summary: Upgrade subscription
      description: Upgrade user's subscription to a higher plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriptionUpgrade'
      responses:
        '200':
          description: Subscription upgraded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        '400':
          description: Invalid upgrade request

  # User Management Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Plan:
      type: object
      properties:
        id:
          type: integer
        plan_name:
          type: string
          enum: [free, pro, business]
        price_monthly:
          type: integer
          description: Price in Japanese Yen
        features:
          type: object
          additionalProperties: true
        api_quota_daily:
          type: integer
        ai_analysis_quota_daily:
          type: integer
        is_active:
          type: boolean
    Subscription:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        plan_id:
          type: integer
        status:
          type: string
          enum: [active, inactive, cancelled, expired]
        current_period_start:
          type: string
          format: date-time
        current_period_end:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    SubscriptionWithUsage:
      allOf:
        - $ref: '#/components/schemas/Subscription'
        - type: object
          properties:
            plan:
              $ref: '#/components/schemas/Plan'
            usage_quota:
              type: object
              properties:
                api_usage_today:
                  type: integer
                ai_analysis_usage_today:
                  type: integer
                api_quota_daily:
                  type: integer
                ai_analysis_quota_daily:
                  type: integer
    SubscriptionUpgrade:
      type: object
      required:
        - plan_id
      properties:
        plan_id:
          type: integer
          description: Target plan ID to upgrade to

    # Response Schemas
...